// src/app/api/model/company-data-confirm-response.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CompanyType } from './company-type';


export interface CompanyDataConfirmResponse { 
    /**
     * AccountStatus enum name
     */
    accountStatus?: CompanyDataConfirmResponseAccountStatusEnum;
    activated?: boolean;
    companyDataId?: number;
    companyName?: string;
    companyType?: CompanyType;
    message?: string;
    nip?: string;
}
export enum CompanyDataConfirmResponseAccountStatusEnum {
    INACTIVE = 'INACTIVE',
    IN_VALIDATION = 'IN_VALIDATION',
    ACTIVE = 'ACTIVE',
    BANNED = 'BANNED',
    BLOCKED_DUE_TO_NOT_ACCEPTING_TERMS = 'BLOCKED_DUE_TO_NOT_ACCEPTING_TERMS',
    TO_BE_DELETED = 'TO_BE_DELETED',
    DELETED = 'DELETED'
};